Wait between commits in auditing purge test to allow system clock to change
git-svn-id: http://repo.aduna-software.org/svn/org.openrdf/alibaba/trunk@11482 3fbd4d3e-0f96-47c4-bab7-89622e02a19e

+++ auditing-sail/src/main/java/org/openrdf/sail/auditing/AuditingSail.java
@@ -229,9 +229,9 @@ public class AuditingSail extends SailWrapper {
 			this.predecessors.removeAll(set);
 			this.predecessors.add(trx);
 		}
-		if (purgeAfter != null) {
+		if (purgeAfter != null && nextPurge < Long.MAX_VALUE) {
 			long now = System.currentTimeMillis();
-			if (nextPurge < Long.MAX_VALUE && now >= nextPurge) {
+			if (now >= nextPurge) {
 				SailConnection con = super.getConnection();
 				try {
 					purgeObsolete(now, con);
